package/apache: add option to disable server
Other packages (e.g. mongoose) can use htdigest and htpasswd, but those
are only available with apache.

We don't want to build the whole apache server just for those tools, so
we add an option to disable the server; it is enabled by default for
legacy purposes (so that existing (def)configs still work).

However, there is no way to tell the apache buildsystem to only build
those two tools, so we have to provide custom build and install
commands.

Cc: Jim Reinhart <jimr@tekvox.com>
Cc: James Autry <jautry@tekvox.com>
Cc: Matthew Maron <matthewm@tekvox.com>
Signed-off-by: Giulio Benetti <giulio.benetti+tekvox@benettiengineering.com>